+++ libft/libft_printf/sources/handlers/handle_p.c
@@ -6,7 +6,7 @@
 /*   By: qcharpen <marvin@le-101.fr>                +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/09/25 06:55:25 by qcharpen     #+#   ##    ##    #+#       */
-/*   Updated: 2018/12/12 07:31:31 by gmadec      ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/11/17 14:19:24 by gmadec      ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -51,10 +51,10 @@ t_list	*ftprintf_handle_p(t_flags *spec, va_list args)
 	tmp = ft_itoa_base_unsigned((long int)arg, 16, 0);
 	if (spec->width == -1)
 		spec->width = 0;
-	len = MAX(spec->prec + 3, MAX(ft_strlen(tmp) + 3, spec->width + 1));
+	len = MAX(spec->prec + 3, MAX((int)ft_strlen(tmp) + 3, spec->width + 1));
 	rst = ft_memalloc(sizeof(*tmp) * len);
 	i = ft_tabset(2);
-	while (!spec->flags[minus] && spec->width > (i[0] + ft_strlen(tmp) + 2))
+	while (!spec->flags[minus] && spec->width > i[0] + (int)ft_strlen(tmp) + 2)
 		rst[(i[0])++] = ' ';
 	rst[(i[0])++] = '0';
 	rst[(i[0])++] = 'x';
